requirements.txt
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
arrow==1.4.0
auto-py-to-exe==2.48.1
binaryornot==0.4.4
blinker==1.9.0
bottle==0.13.4
bottle-websocket==0.2.9
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cookiecutter==2.6.0
Eel==0.18.2
fastapi==0.128.0
fcalc==1.0.3
Flask==3.1.2
flet==0.80.0
flet-cli==0.80.0
flet-desktop==0.80.0
flet-web==0.80.0
future==1.0.0
gevent==25.9.1
gevent-websocket==0.10.1
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
numpy==2.4.0
oauthlib==3.3.1
packaging==25.0
pefile==2024.8.26
Pillow>=10.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.11
pyparsing==3.3.1
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pywin32-ctypes==0.2.3
PyYAML==6.0.3
qrcode==8.2
repath==0.9.0
requests==2.32.5
rich==14.2.0
setuptools==80.9.0
shiboken6==6.10.1
six==1.17.0
starlette==0.50.0
text-unidecode==1.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.40.0
watchdog==6.0.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.4
zope.event==6.1
zope.interface==8.1.1
